requirements/base.txt
Django==2.2.2
django-cors-headers==3.0.2
djangorestframework==3.9.4
psycopg2==2.7.4
pytz==2019.1
sqlparse==0.3.0
